requirements.txt
OttomanAgentAssets[pinecone] @ git+https://github.com/OttomanAI/AgentAssets.git
python-dotenv
